+++ processor/src/test/fixtures/expected/com/example/component_name_ref/Arez_ComponentNameModel.java
@@ -40,6 +40,8 @@ public final class Arez_ComponentNameModel extends ComponentNameModel implements
     return $$arezi$$_id();
   }
 
+  @Override
+  @Nonnull
   public final String getComponentName() {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'getComponentName' invoked on uninitialized component of type 'ComponentNameModel'" );