+++ processor/src/test/fixtures/expected/com/example/context_ref/Arez_SimpleComponent.java
@@ -30,6 +30,7 @@ final class Arez_SimpleComponent extends SimpleComponent implements Disposable, 
   }
 
   @Override
+  @Nonnull
   final ArezContext getContext() {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'getContext' invoked on uninitialized component of type 'SimpleComponent'" );